chore: debug only stuff

+++ examples/triangle/triangle.odin
@@ -737,20 +737,26 @@ cleanup :: proc(s: ^State, data: ^Render_Data) {
 }
 
 main :: proc() {
-	context.logger = log.create_console_logger(opt = {.Level, .Terminal_Color})
-	defer log.destroy_console_logger(context.logger)
+	when ODIN_DEBUG {
+		context.logger = log.create_console_logger(opt = {.Level, .Terminal_Color})
+		defer log.destroy_console_logger(context.logger)
 
-	track: mem.Tracking_Allocator
-	mem.tracking_allocator_init(&track, context.allocator)
-	context.allocator = mem.tracking_allocator(&track)
-	defer mem.tracking_allocator_destroy(&track)
+		track: mem.Tracking_Allocator
+		mem.tracking_allocator_init(&track, context.allocator)
+		context.allocator = mem.tracking_allocator(&track)
+		defer mem.tracking_allocator_destroy(&track)
 
-	defer {
-		for _, leak in track.allocation_map {
-			fmt.printf("%v leaked %v bytes\n", leak.location, leak.size)
-		}
-		for bad_free in track.bad_free_array {
-			fmt.printf("%v allocation %p was freed badly\n", bad_free.location, bad_free.memory)
+		defer {
+			for _, leak in track.allocation_map {
+				fmt.printf("%v leaked %v bytes\n", leak.location, leak.size)
+			}
+			for bad_free in track.bad_free_array {
+				fmt.printf(
+					"%v allocation %p was freed badly\n",
+					bad_free.location,
+					bad_free.memory,
+				)
+			}
 		}
 	}
